+++ rules/doctrine/tests/Rector/ClassMethod/ServiceEntityRepositoryParentCallToDIRector/Source/Project.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Doctrine\Tests\Rector\ClassMethod\ServiceEntityRepositoryParentCallToDIRector\Source;
+
+class Project
+{
+
+}
